ci: Bundle numbered integration tests to speed up link time in tests
