+++ crates/edr_provider/tests/issues/issue_346.rs
@@ -1,5 +1,3 @@
-#![cfg(feature = "test-utils")]
-
 use edr_provider::{test_utils::create_test_config, time::CurrentTime, NoopLogger, Provider};
 use serde_json::json;
 use tokio::runtime;
